Modularize program pt. 1 (extremely buggy)
Grouped Superstore-unique traits into Superstore class while
keeping the main bulk of the program generic. The hope is to
create multiple Store classes which can be easily fed into
the program. Unfortunately in the migration, many bugs
were introduced. To name a few:

-on webdriver, Superstore sometimes has a "sponsored" badge
which happens to share the exact same class name as the
sale price, ruining the program. This happened to the Condensed Milk.

-item[6] isn't being counted any more, not sure why. Regular price
has a full null column and is still being counted as empty strings,
while cheapest price is returning an index error. In addition,
using a seemingly acceptable try except resolves the index error,
but item[6] still doesn't get counted as "".

-changed most, if not all, XPATHs to class names or css selectors,
for the sake of futureproofing. some of the new selectors might
be pointing at the wrong elements now as a result.

-if an item has a sale div, BUT it's still above
the max buying price, the logic tree is bugged. ex. won't print
out "Item not on sale", probably other bugs. see: Annie's Crackers